fix(keystore): the TA passed a uint32_t where the API writes a size_t
TEE_GetObjectBufferAttribute takes a size_t *, and cmd_export_once handed
it the address of a uint32_t. On aarch64 that is an incompatible pointer
type and, if a compiler allowed it, eight bytes written into a four-byte
object.

The cause is that OP-TEE 4.x defaults to the GlobalPlatform 1.2 types,
where TEE_Param.memref.size is size_t. The 1.1 types with uint32_t are
still available behind CFG_TA_OPTEE_CORE_API_COMPAT_1_1, which the OP-TEE
example TAs set in their makefiles and this one does not, so code adapted
from an example compiles there and not here.

Found by reading the headers rather than by a build, because nothing
builds this: CI compiles bench-status.c, lte-gpio.c and rt-toggle.c, and
none of the three C files in this project, since two need libteec and the
third needs the TA development kit. The green tick on the previous commit
covered the shell, the Python and the lint and said nothing about the C.

Checked at the same time and sound: bool is available through
tee_api_types.h, the persistent-object calls take size_t lengths and get
strlen and sizeof, TEE_MACComputeFinal is handed memref.size and its
address, and the client side is clean because TEEC_TempMemoryReference.size
is size_t throughout.

The repair for the class rather than the instance is a CI step compiling
benchkey.c against libteec headers from a pinned tag, the same shape as
the existing job that builds libgpiod v2 from a pinned tag for the GPIO
daemon. Not in this commit: that is a change to the shared workflow.

+++ meta-bench/recipes-bench/bench-keystore/files/bench_keystore_ta.c
@@ -185,7 +185,22 @@ static TEE_Result cmd_export_once(uint32_t types, TEE_Param params[4])
 						TEE_PARAM_TYPE_NONE);
 	TEE_ObjectHandle key = TEE_HANDLE_NULL;
 	TEE_ObjectHandle lock = TEE_HANDLE_NULL;
-	uint32_t size;
+	/*
+	 * size_t, not uint32_t, and the difference is not cosmetic.
+	 *
+	 * In OP-TEE 4.x the GlobalPlatform 1.2 types are the default, so
+	 * TEE_Param.memref.size is size_t and TEE_GetObjectBufferAttribute
+	 * takes a size_t *. On aarch64 that is eight bytes. Passing the
+	 * address of a uint32_t is an incompatible pointer type, and if a
+	 * compiler let it through the callee would write eight bytes into
+	 * a four-byte object and corrupt the stack next to it.
+	 *
+	 * The 1.1 types, with uint32_t, are still available behind
+	 * CFG_TA_OPTEE_CORE_API_COMPAT_1_1, which the examples set and
+	 * this TA does not. Code copied from an example built that way
+	 * compiles there and not here.
+	 */
+	size_t size;
 	TEE_Result res;
 
 	if (types != expect)
